store: improve unsupported schema version error message
