+++ graph/src/components/store/err.rs
@@ -1,9 +1,8 @@
+use super::{BlockNumber, DeploymentHash, DeploymentSchemaVersion};
+use crate::prelude::QueryExecutionError;
 use anyhow::{anyhow, Error};
 use thiserror::Error;
 use tokio::task::JoinError;
-
-use super::{BlockNumber, DeploymentHash};
-use crate::prelude::QueryExecutionError;
 
 #[derive(Error, Debug)]
 pub enum StoreError {
@@ -51,7 +50,9 @@ pub enum StoreError {
     #[error("panic in subgraph writer: {0}")]
     WriterPanic(JoinError),
     #[error(
-        "found deployment schema version {0} which is not supported. Did you downgrade Graph Node?"
+        "found schema version {0} but this graph node only supports versions up to {}. \
+         Did you downgrade Graph Node?",
+        DeploymentSchemaVersion::LATEST
     )]
     UnsupportedDeploymentSchemaVersion(i32),
 }
